moved phpDocumentor/Command to phpDocumentor/Application/Cli/Command
