+++ src/phpDocumentor/Application/Cli/Command/Command.php
@@ -9,7 +9,7 @@
  * @link      http://phpdoc.org
  */
 
-namespace phpDocumentor\Command;
+namespace phpDocumentor\Application\Cli\Command;
 
 use Symfony\Component\Console\Command\Command as SymfonyCommand;
 use Symfony\Component\Console\Helper\HelperSet;
