+++ src/phpDocumentor/Application/Cli/Command/RunCommand.php
@@ -16,8 +16,7 @@ namespace phpDocumentor\Application\Cli\Command;
 use Desarrolla2\Cache\Adapter\AdapterInterface;
 use Desarrolla2\Cache\Adapter\File;
 use Desarrolla2\Cache\CacheInterface;
-use phpDocumentor\Command\Command;
-use phpDocumentor\Command\Helper\ConfigurationHelper;
+use phpDocumentor\Application\Cli\Command\Helper\ConfigurationHelper;
 use phpDocumentor\Compiler\Compiler;
 use phpDocumentor\Compiler\CompilerPassInterface;
 use phpDocumentor\Configuration;
