+++ src/phpDocumentor/ContainerDefinitions.php
@@ -15,9 +15,9 @@ use League\Tactician\Handler\Locator\HandlerLocator;
 use League\Tactician\Handler\MethodNameInflector\InvokeInflector;
 use League\Tactician\Handler\MethodNameInflector\MethodNameInflector;
 use phpDocumentor\Application\Commands\ContainerLocator;
-use phpDocumentor\Command\Helper\ConfigurationHelper;
-use phpDocumentor\Command\Helper\LoggerHelper;
-use phpDocumentor\Command\Phar\UpdateCommand;
+use phpDocumentor\Application\Cli\Command\Helper\ConfigurationHelper;
+use phpDocumentor\Application\Cli\Command\Helper\LoggerHelper;
+use phpDocumentor\Application\Cli\Command\Phar\UpdateCommand;
 use phpDocumentor\Application\Cli\Command\RunCommand;
 use phpDocumentor\Compiler\Compiler;
 use phpDocumentor\Compiler\Linker\Linker;
