+++ src/phpDocumentor/Transformer/Command/Template/ListCommand.php
@@ -11,7 +11,7 @@
 
 namespace phpDocumentor\Transformer\Command\Template;
 
-use phpDocumentor\Command\Command;
+use phpDocumentor\Application\Cli\Command\Command;
 use phpDocumentor\Transformer\Template\Factory;
 use \Symfony\Component\Console\Input\InputInterface;
 use \Symfony\Component\Console\Output\OutputInterface;
